move routefiles out of routes/api to fix structure, update server file
