+++ routes/logout.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const logoutController = require("../../controllers/logoutController");
+const logoutController = require("../controllers/logoutController");
 
 router.route("/").post(logoutController.handleLogout);
