+++ routes/register.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const registerController = require("../../controllers/registerController");
+const registerController = require("../controllers/registerController");
 
 router.route("/").post(registerController.handleNewUser);
